+++ javascripts/events.js
@@ -1,5 +1,5 @@
 const message = require ('./messages/message_main');
-const events = require('./events/main');
+// const events = require('./events/events_main');
 
 const authEvents = () =>
 {
@@ -55,33 +55,33 @@ const authEvents = () =>
   });
 };
 
-const eventsArray = [
-  {
-    'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
-    'event': 'Tug-O-War',
-    'startDate': 1528763298535,
-    'location': 'Woods',
-  },
-  {
-    'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
-    'event': 'Tug-O-War',
-    'startDate': 1528763298535,
-    'location': 'Woods',
-  },
-  {
-    'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
-    'event': 'Tug-O-War',
-    'startDate': 1528763298535,
-    'location': 'Woods',
-  },
-];
+// const eventsArray = [
+//   {
+//     'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
+//     'event': 'Tug-O-War',
+//     'startDate': 1528763298535,
+//     'location': 'Woods',
+//   },
+//   {
+//     'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
+//     'event': 'Tug-O-War',
+//     'startDate': 1528763298535,
+//     'location': 'Woods',
+//   },
+//   {
+//     'userUid': '5ykBb0xyadPZLgH4EPO4i88HIql2',
+//     'event': 'Tug-O-War',
+//     'startDate': 1528763298535,
+//     'location': 'Woods',
+//   },
+// ];
 
 const initializer = () =>
 {
   // GET DATA FROM FIREBASE AND SET MESSAGE BOARD LISTENERS
   message.initMessageBoard();
   authEvents();
-  events.eventsAdded(eventsArray);
+  // events.eventsAdded(eventsArray);
 };
 
 module.exports =
