+++ javascripts/events/events_firebase.js
@@ -1,0 +1,33 @@
+let firebaseConfig = {};
+let uid = '';
+
+const setConfig = (fbConfig) => {
+  firebaseConfig = fbConfig;
+};
+
+const setUID = (newUID) => {
+  uid = newUID;
+};
+
+const getSavedEvents = (whatILike) => {
+  whatILike.uid = uid;
+  return new Promise((resolve, reject) => {
+    $.ajax({
+      method: 'POST',
+      url: `${firebaseConfig.databaseURL}/events.json`,
+      data: JSON.stringify(whatILike),
+    })
+      .done((uniqueKey) => {
+        resolve(uniqueKey);
+      })
+      .fail((err) => {
+        reject(err);
+      });
+  });
+};
+
+module.exports = {
+  setUID,
+  setConfig,
+  getSavedEvents,
+};
